backend/pyproject.toml
[project]
name = "tai-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # Web framework + ASGI server
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",

    # Async PostgreSQL driver used by the app at runtime
    "asyncpg>=0.30",

    # Config — reads .env files and validates env vars as typed fields
    "pydantic-settings>=2.6",

    # Loads .env into the environment before pydantic-settings reads it
    "python-dotenv>=1.0",

    # S3-compatible client for Cloudflare R2 uploads
    "boto3>=1.35",

    # Required by FastAPI to parse multipart/form-data (file uploads)
    "python-multipart>=0.0.12",

    # OpenAI-compatible client — used for OpenRouter API calls
    "openai>=1.0",

    # RAG ingestion pipeline
    "llama-index-core>=0.11",
    "llama-index-embeddings-huggingface>=0.3",
    "pypdf>=4.0",
    "python-docx>=1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",   # async HTTP client for testing FastAPI routes
    "ruff>=0.8",     # linter + formatter
]

# We run uvicorn directly from inside backend/ so Python finds the local folders
# (core/, db/, routers/ etc.) automatically via the working directory.
# No packages need to be installed — this section just stops setuptools from
# trying to auto-discover them and erroring out.
[tool.setuptools]
packages = []
